add_booking method now uses user_id and space_id

+++ lib/bookings.rb
@@ -6,19 +6,19 @@ class Bookings
     date_array_formatted = date_array.map {|date| date.strftime("%d-%m-%Y")}
   end
 
-  def self.check_availability(dates = [])
+  def self.check_availability(dates = [], space_id)
     dates.map do |date|
-      result = DatabaseConnection.query("SELECT space_id, date FROM bookings WHERE date = $1 AND approved = $2;", [date, true])
+      result = DatabaseConnection.query("SELECT space_id, date FROM bookings WHERE date = $1 AND approved = $2 AND space_id = $3;", [date, true, space_id])
       result.count.zero? ? "Available" : "Not Available"
     end
   end
 
-  def self.add_booking(date)
-    DatabaseConnection.query("INSERT INTO bookings(date) VALUES($1);", [date])
+  def self.add_booking(date, space_id, user_id)
+    DatabaseConnection.query("INSERT INTO bookings(date, space_id, user_id) VALUES($1, $2, $3);", [date, space_id, user_id])
   end
 
-  def self.remove_booking(date)
-    DatabaseConnection.query("DELETE FROM bookings WHERE date = $1;", [date])
+  def self.remove_booking(date, space_id, user_id)
+    DatabaseConnection.query("DELETE FROM bookings WHERE date = $1 AND space_id = $2 AND user_id = $3;", [date, space_id, user_id])
   end
